+++ packages/types/custom/core.ts
@@ -599,6 +599,11 @@ export interface PolarMapOptions {
   zoomLevel: number
 }
 
+export interface AuthenticationConfig {
+  interceptorUrlRegex: string
+  tokenName: string
+}
+
 /** The initial language the client should be using; defaults to 'de' if not given */
 export type InitialLanguage = 'de' | 'en'
 
@@ -658,11 +663,11 @@ export interface MasterportalApiConfig {
 export interface MapConfig extends MasterportalApiConfig {
   /** Configured layers */
   layers: LayerConfiguration[]
+  authentication?: AuthenticationConfig
   /** if true, all services' availability will be checked with head requests */
   checkServiceAvailability?: boolean
   extendedMasterportalapiMarkers?: ExtendedMasterportalapiMarkers
   featureStyles?: string
-  interceptorUrlRegex?: string
   language?: InitialLanguage
   locales?: Locale[]
   renderFaToLightDom?: boolean