+++ packages/plugins/Zoom/src/types.ts
@@ -1,4 +1,5 @@
-import { ZoomIcons } from '@polar/lib-custom-types'
+import { ZoomConfiguration, ZoomIcons } from '@polar/lib-custom-types'
+import { VueConstructor } from 'vue'
 
 export interface ZoomState {
   zoomLevel: number
@@ -7,6 +8,8 @@ export interface ZoomState {
 }
 
 export interface ZoomGetters extends ZoomState {
+  configuration: ZoomConfiguration
+  component: VueConstructor | null
   icons: ZoomIcons
   maximumZoomLevelActive: boolean
   minimumZoomLevelActive: boolean