+++ packages/clients/diplan/src/store/geoEditing/cutPolygons/types.ts
@@ -1,0 +1,5 @@
+import { WriteOptions } from 'ol/format/Feature'
+
+export type ProjectionInfo = Required<
+  Pick<WriteOptions, 'dataProjection' | 'featureProjection'>
+>