+++ packages/types/custom/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Feature: Add `snapTo` to `DrawConfiguration` for specification of layers to snap to.
 - Feature: Add `lassos` to `DrawConfiguration`. With this, `addLoading`, `removeLoading`, and `toastAction` have also been introduced to allow the feature to use other plugins via API calls, and `Lasso` itself has been introduced.
-- Feature: Add new configuration parameter `interceptorUrlRegex` to `MapConfig`.
+- Feature: Add new optional configuration parameter `authentication` to `MapConfig`.
+- Feature: Add new type `AuthenticationConfiguration`.
 
 ## 2.0.0